Add composer
