+++ phpunit.xml
@@ -1,0 +1,17 @@
+<phpunit backupGlobals="false"
+         backupStaticAttributes="false"
+         bootstrap="vendor/autoload.php"
+         colors="true"
+         convertErrorsToExceptions="true"
+         convertNoticesToExceptions="true"
+         convertWarningsToExceptions="true"
+         processIsolation="false"
+         stopOnFailure="false"
+>
+    <testsuites>
+        <testsuite name="Application Test Suite">
+            <file>./tests/e2e/Client.php</file>
+            <directory>./tests/</directory>
+        </testsuite>
+    </testsuites>
+</phpunit>
